+++ vanity/db/list_database.h
@@ -43,12 +43,12 @@ public:
 	// get the value for a list key at a given index
 	// returns the value, or ErrorKind::NotList if the value is not a list
 	// or ErrorKind::OutOfRange if the index is out of range or if value does not exist
-	std::variant<std::string, ErrorKind> list_get(const key_type &key, int index);
+	std::variant<std::string, ErrorKind> list_get(const key_type &key, int64_t index);
 
 	// set the value for a list key at a given index
 	// returns true if the value was set, or ErrorKind::NotList if the value is not a list
 	// or ErrorKind::OutOfRange if the index is out of range or if value does not exist
-	std::variant<bool, ErrorKind> list_set(const key_type &key, int index, const std::string& value);
+	std::variant<bool, ErrorKind> list_set(const key_type &key, int64_t index, const std::string& value);
 
 	// push values to the top of a list
 	// returns the new length, or ErrorKind::NotList if the value is not a list
@@ -62,53 +62,53 @@ public:
 	// returns the values or ErrorKind::NotList if the value is not a list
 	// if the index is positive and out of range, returns the values up to the end of the list
 	// if the index is negative and out of range, returns an empty list
-	std::variant<list_t, ErrorKind> list_pop_left(const key_type &key, int n);
+	std::variant<list_t, ErrorKind> list_pop_left(const key_type &key, int64_t n);
 
 	// pop n values from the bottom of a list
 	// returns the values, or ErrorKind::NotList if the value is not a list
 	// if the index is positive and out of range, returns the values up to the end of the list
 	// if the index is negative and out of range, returns an empty list
-	std::variant<list_t, ErrorKind> list_pop_right(const key_type &key, int n);
+	std::variant<list_t, ErrorKind> list_pop_right(const key_type &key, int64_t n);
 
 	// get the value for a range of a list key inclusively
 	// returns the values, or ErrorKind::NotList if the value is not a list
 	// returns an empty list if the range is out of bounds or invalid
-	std::variant<list_t, ErrorKind> list_range(const key_type &key, int start, int end);
+	std::variant<list_t, ErrorKind> list_range(const key_type &key, int64_t start, int64_t end);
 
 	// trim the list stored at key, so that it will
 	// contain only the specified range of elements (inclusively)
 	// returns the number of trimmed elements
 	// or ErrorKind::NotList if the value is not a list
-	std::variant<size_t, ErrorKind> list_trim(const key_type &key, int start, int end);
+	std::variant<size_t, ErrorKind> list_trim(const key_type &key, int64_t start, int64_t end);
 
 	// remove a number of elements equal to count from the list stored at key
 	// that hold the value element
 	// returns the number of removed elements, or ErrorKind::NotList if the value is not a list
 	// removes from the end if count is negative or all elements if count is 0
-	std::variant<size_t, ErrorKind> list_remove(const key_type &key, const std::string& element, int count);
+	std::variant<size_t, ErrorKind> list_remove(const key_type &key, const std::string& element, int64_t count);
 
 private:
 	// get the iterator for a list key at a given index
 	// returns the iterator, or ErrorKind::NotList if the value is not a list
 	// or ErrorKind::OutOfRange if the index is out of range or if value does not exist
 	// index can be negative to get the element from the end of the list
-	std::variant<list_t::iterator, ErrorKind> iterator_or_error(const key_type &key, int index);
+	std::variant<list_t::iterator, ErrorKind> iterator_or_error(const key_type &key, int64_t index);
 
 	// get the iterator for a given list at a given index
 	// returns the iterator, or the end iterator if the index is out of range
 	// index can be negative to get the element from the end of the list
-	static list_t::iterator iterator_or_end(list_t &list, int index);
+	static list_t::iterator iterator_or_end(list_t &list, int64_t index);
 
 	// get the reverse iterator for a given list at a given index
 	// returns the iterator, or the rend iterator if the index is out of range
 	// index can be negative to get the element from the end of the list
-	static list_t::reverse_iterator reverse_iterator_or_rend(list_t &list, int index);
+	static list_t::reverse_iterator reverse_iterator_or_rend(list_t &list, int64_t index);
 
 	// get a pair of iterators for a given list at given indexes inclusively
-	static std::pair<list_t::iterator, list_t::iterator> iterator_pair_inclusive(list_t &list, int start, int end);
+	static std::pair<list_t::iterator, list_t::iterator> iterator_pair_inclusive(list_t &list, int64_t start, int64_t end);
 
 	// check if a pair of integers form an invalid range
-	static bool is_invalid_range(int start, int end);
+	static bool is_invalid_range(int64_t start, int64_t end);
 };
 
 } // namespace vanity::db
